+++ client/src/components/register/UploadDoneBtn.js
@@ -26,7 +26,7 @@ function UploadDoneBtn(props) {
 
   async function registerCalculet() {
     try {
-      await axios.post("/calculets", props).then((response) => {
+      await axios.post("/calculets/", props).then((response) => {
         console.log("계산기 등록 완료!");
         console.log(response.data);
         console.log(props);
