--- server/routes/temp.js
@@ -1,1 +1,0 @@
-// 폴더 업로드 용
